+++ src/main/java/no/taardal/pixelcave/state/actor/ActorIdleState.java
@@ -2,7 +2,7 @@ package no.taardal.pixelcave.state.actor;
 
 import no.taardal.pixelcave.actor.Actor;
 import no.taardal.pixelcave.statemachine.StateMachine;
-import no.taardal.pixelcave.vector.Vector2d;
+import no.taardal.pixelcave.vector.Vector2f;
 
 public abstract class ActorIdleState<T extends Actor> extends ActorState<T> {
 
@@ -12,12 +12,12 @@ public abstract class ActorIdleState<T extends Actor> extends ActorState<T> {
 
     @Override
     public void onEntry() {
-        actor.setVelocity(Vector2d.zero());
+        actor.setVelocity(Vector2f.zero());
         updateBounds();
     }
 
     @Override
-    public void update(double secondsSinceLastUpdate) {
+    public void update(float secondsSinceLastUpdate) {
         getAnimation().update();
     }
