+++ src/main/java/no/taardal/pixelcave/state/game/GameState.java
@@ -6,7 +6,7 @@ import no.taardal.pixelcave.keyboard.Keyboard;
 public interface GameState {
 
     GameState handleInput(Keyboard keyboard);
-    GameState update(double secondsSinceLastUpdate, Camera camera);
+    GameState update(float secondsSinceLastUpdate, Camera camera);
     void draw(Camera camera);
 
 }
